refactor: 약관 모듈을 역할별 디렉토리로 분리
입력 검증은 termMiddleware로 이동하고, 타입·상수는 types/constants로 분리해 controller는 요청/응답 처리에 집중하도록 정리했습니다.
