+++ src/constants/termConstants.ts
@@ -1,0 +1,1 @@
+export const TERM_TYPES = ['REQUIRED', 'OPTIONAL'] as const;
