+++ src/types/term.types.ts
@@ -1,0 +1,11 @@
+import type { TERM_TYPES } from '../constants/termConstants.js';
+
+export type TermType = typeof TERM_TYPES[number];
+
+export type GetTermsQuery = {
+  type?: TermType;
+};
+
+export interface AgreeTermsRequestBody {
+  termIds: string[];
+}
